@@ -3,9 +3,9 @@
   "description": "A template for start coding React projects faster.",
   "license": "MIT",
   "main": "index.js",
-  "name": "react-projetc-template",
+  "name": "react-project-template",
   "private": true,
-  "repository": "https://github.com/TituxMetal/react-projetc-template.git",
+  "repository": "https://github.com/TituxMetal/react-project-template.git",
   "version": "1.0.0",
   "scripts": {
     "clear": "rm -rf ./.parcel-cache ./.cache ./dist",
@@ -15,22 +15,22 @@
   },
   "dependencies": {
     "@fontsource/roboto": "^4.5.1",
-    "@hookform/resolvers": "^2.8.2",
+    "@hookform/resolvers": "^2.8.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-hook-form": "^7.17.5",
-    "react-router": "^5.2.1",
-    "react-router-dom": "^5.3.0",
+    "react-hook-form": "^7.20.2",
+    "react-router": "^6.0.2",
+    "react-router-dom": "^6.0.2",
     "styled-components": "^5.3.3",
     "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.8",
-    "@babel/plugin-transform-react-jsx": "^7.14.9",
+    "@babel/core": "^7.16.0",
+    "@babel/plugin-transform-react-jsx": "^7.16.0",
     "@tailwindcss/typography": "^0.4.1",
-    "autoprefixer": "^10.3.7",
+    "autoprefixer": "^10.4.0",
     "env-cmd": "^10.1.0",
-    "eslint": "^8.1.0",
+    "eslint": "^8.3.0",
     "eslint-config-kentcdodds": "^19.2.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-import-resolver-alias": "^1.1.2",
@@ -38,7 +38,7 @@
     "postcss": "^8.3.11",
     "prettier": "^2.4.1",
     "tailwindcss": "^2.2.17",
-    "twin.macro": "^2.8.0"
+    "twin.macro": "^2.8.1"
   },
   "alias": {
     "~": "./src"
