@@ -2,12 +2,12 @@ import React from 'react'
 import { render } from 'react-dom'
 
 import { GlobalStyle } from '~/core'
-import Routes from '~/routes'
+import Routing from '~/routes'
 
 const element = document.getElementById('root')
 render(
   <GlobalStyle>
-    <Routes />
+    <Routing />
   </GlobalStyle>,
   element
 )
